Use mono.cecil to load symbol files and only try alternative method if that files
