+++ main/OpenCover.Framework/OpenCover.Framework.csproj
@@ -153,6 +153,7 @@
     <Compile Include="Symbols\CecilSymbolManager.cs" />
     <Compile Include="Symbols\ISymbolManager.cs" />
     <Compile Include="Symbols\SymbolFile.cs" />
+    <Compile Include="Symbols\SymbolFileHelper.cs" />
     <Compile Include="Utility\CodeCoverageStringTextSource.cs" />
     <Compile Include="Utility\IPerfCounters.cs" />
     <Compile Include="Utility\LogHelper.cs" />
