+++ main/OpenCover.Test/App.config
@@ -10,10 +10,10 @@
   </autofac>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
+      <!--<dependentAssembly>
         <assemblyIdentity name="Mono.Cecil" publicKeyToken="0738eb9f132ed756" culture="neutral" />
         <bindingRedirect oldVersion="0.0.0.0-0.10.0.0" newVersion="0.10.0.0" />
-      </dependentAssembly>
+      </dependentAssembly>-->
       <dependentAssembly>
         <assemblyIdentity name="Autofac" publicKeyToken="17863af14b0044da" culture="neutral" />
         <bindingRedirect oldVersion="0.0.0.0-3.5.0.0" newVersion="3.5.0.0" />
