+++ main/OpenCover.Test/OpenCover.Test.csproj
@@ -185,8 +185,11 @@
     <Compile Include="Extensions\Strategy\TrackNUnitTestMethodsTests.cs" />
     <Compile Include="Extensions\Strategy\TrackXUnitTestMethodsTests.cs" />
     <Compile Include="Framework\Strategy\TrackedMethodStrategyManagerTests.cs" />
+    <Compile Include="Framework\Symbols\BaseMdbTests.cs" />
     <Compile Include="Framework\Symbols\CecilSymbolManagerMdbTests.cs" />
     <Compile Include="Framework\Symbols\CecilSymbolManagerTests.cs" />
+    <Compile Include="Framework\Symbols\SymbolFileHelperMdbTests.cs" />
+    <Compile Include="Framework\Symbols\SymbolFileHelperTests.cs" />
     <Compile Include="Framework\Utility\PerfCountersTests.cs" />
     <Compile Include="Framework\Utility\CodeCoverageStringTextSourceTest.cs" />
     <Compile Include="Framework\Utility\SequencePointComparerTest.cs" />
